feat: extract database abstraction layer to harness-core
Move open_pool, DbEntity, Db<T> from harness-server into
harness-core::db so all crates share a single implementation.
Add Migration/Migrator for versioned schema evolution.

- harness-core/src/db.rs: new module with open_pool, DbEntity,
  Db<T>, Migration, Migrator (with duplicate-column tolerance)
- harness-server/src/db.rs: re-exports from harness_core::db
- harness-server/src/task_db.rs: use Migrator with 4 versioned
  migrations (baseline + source/external_id/parent_id columns)
- harness-server/src/thread_db.rs: use Migrator
- harness-exec/src/plan.rs: impl DbEntity for ExecPlan (orphan fix)
- harness-observe/src/event_store.rs: use shared open_pool and
  Migrator, eliminates duplicated pool creation code
- Bump version to 0.6.4

Closes #229
