+++ crates/harness-core/Cargo.toml
@@ -14,6 +14,7 @@ thiserror = { workspace = true }
 async-trait = { workspace = true }
 tokio = { workspace = true, features = ["sync"] }
 anyhow = { workspace = true }
+sqlx = { workspace = true }
 
 [dev-dependencies]
 anyhow = { workspace = true }
