+++ crates/harness-core/src/lib.rs
@@ -2,6 +2,7 @@
 pub mod agent;
 pub mod agents_md;
 pub mod config;
+pub mod db;
 pub mod error;
 pub mod interceptor;
 pub mod lang_detect;
